ducky/storage: move listdir to storage
